+++ src/lambda/digitalChannelHealthCheck/index.js
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: MIT-0
 const { log } = require('common-util');
 const fb = require('./lib/facebook');
+const wa = require('./lib/whatsapp');
 
 exports.handler = async (event) => {
   log.debug('Event', event);
@@ -9,7 +10,10 @@ exports.handler = async (event) => {
   switch (event.rawPath) {
     case '/webhook/facebook':
       log.debug('Facebook channel detected.');
-      return fb.handler(event);
+      return await fb.handler(event);
+    case '/webhook/whatsapp':
+      log.debug('WhatsApp channel detected.');
+      return await wa.handler(event);
     default:
       log.warn(
         `Request path "${event.rawPath}" does not match any expected paths.`